third_party/blink/web_tests/media/controls/accessibility-download-button.html

<!DOCTYPE html>
<title>Media Controls: download button accessibility tests</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/run-after-layout-and-paint.js"></script>
<script src="../media-controls.js"></script>
<video controls preload="none" src="https://someexample.example/example.mp4"></video>
<script>
async_test(t => {
  var video = document.querySelector('video');

  runAfterLayoutAndPaint(t.step_func_done(_ => {
    assert_true(isDownloadsButtonEnabled(video));

    var download_button = downloadButton(video);
    var download_overflow_item = downloadsOverflowItem(video);
    assert_not_equals(download_button, null);
    assert_not_equals(download_overflow_item, null);

    assert_equals(download_button.getAttribute('aria-label'),
                  'download media');

    assert_equals(download_overflow_item.getAttribute('aria-label'),
                  'download media');
  }));
});
</script>
